refactor: extract shared time formatter
