+++ src/components/OfflineProgressModal.jsx
@@ -1,15 +1,5 @@
 import { useGame } from '../state/useGame.js';
-
-function formatTime(seconds) {
-  const h = Math.floor(seconds / 3600);
-  const m = Math.floor((seconds % 3600) / 60);
-  const s = seconds % 60;
-  const parts = [];
-  if (h) parts.push(`${h}h`);
-  if (m) parts.push(`${m}m`);
-  if (s || parts.length === 0) parts.push(`${s}s`);
-  return parts.join(' ');
-}
+import { formatTime } from '../utils/time.js';
 
 export default function OfflineProgressModal() {
   const { state, dismissOfflineModal } = useGame();
