+++ src/views/__tests__/BaseView.test.jsx
@@ -83,7 +83,9 @@ function BuildingRow({ building }) {
       <div className="flex items-center justify-between">
         <span>
           {building.name}{' '}
-          {building.maxCount != null ? `${count}/${building.maxCount}` : `(${count})`}
+          {building.maxCount != null
+            ? `${count}/${building.maxCount}`
+            : `(${count})`}
         </span>
         <div className="space-x-2">
           <button
@@ -97,8 +99,8 @@ function BuildingRow({ building }) {
                     building.requiresResearch
                   }`
                 : atMax
-                ? `Max ${building.maxCount}`
-                : undefined
+                  ? `Max ${building.maxCount}`
+                  : undefined
             }
           >
             Build
